% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{cnt_logger_S3_register}
\alias{cnt_logger_S3_register}
\title{Register S3 methods for logger connector}
\usage{
cnt_logger_S3_register()
}
\description{
This function registers S3 methods for reading, writing, and removing
connectors with the logger connector class.
}
