% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.logger-package.R
\docType{package}
\name{connector.logger-package}
\alias{connector.logger}
\alias{connector.logger-package}
\title{connector.logger: What the Package Does (One Line, Title Case)}
\description{
The goal of connector.logger is to implement the log function(s) from whirl to log read, write, and delete operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NN-OpenSource/connector.logger}
  \item \url{https://miniature-couscous-2kqngew.pages.github.io/}
  \item Report bugs at \url{https://github.com/NN-OpenSource/connector.logger/issues}
}

}
\author{
\strong{Maintainer}: Aksel Thomsen \email{oath@novonordisk.com}

Authors:
\itemize{
  \item Lovemore Gakava \email{lvgk@novonordisk.com}
  \item Cervan Girard \email{cgid@novonordisk.com}
  \item Kristian Troejelsgaard \email{ktqn@novonordisk.com}
  \item Steffen Falgreen Larsen \email{sffl@novonordisk.com}
  \item Vladimir Obucina \email{vlob@novonordisk.com}
}

Other contributors:
\itemize{
  \item Novo Nordisk A/S [copyright holder]
}

}
\keyword{internal}
