% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{list_content_cnt.connector_logger}
\alias{list_content_cnt.connector_logger}
\title{List contents Operation for connector_logger class}
\usage{
\method{list_content_cnt}{connector_logger}(connector_object, ...)
}
\arguments{
\item{connector_object}{The connector_logger object.}

\item{...}{Additional parameters.}
}
\value{
The result of the read operation.
}
\description{
Implementation of the log_read_connector function for the connector_logger
class.
}
