% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{list_content_cnt}
\alias{list_content_cnt}
\title{List available content from the connector}
\usage{
list_content_cnt(connector_object, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
A \link{character} vector of content names
}
\description{
Addition list content methods for databricks connectors implemented for
\code{\link[connector:list_content_cnt]{connector::list_content_cnt()}}:
}
