% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_read_connector.connector_dbi}
\alias{log_read_connector.connector_dbi}
\title{Log read operation for connector dbi}
\usage{
\method{log_read_connector}{connector_dbi}(connector_object, name, ...)
}
\description{
Log read operation for connector dbi
}
