% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_read_connector.connector_fs}
\alias{log_read_connector.connector_fs}
\title{Log read operation for FS connector}
\usage{
\method{log_read_connector}{connector_fs}(connector_object, name, ...)
}
\description{
Log read operation for FS connector
}
