% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sharepoint_cnt.R
\name{log_read_connector.Connector_sharepoint}
\alias{log_read_connector.Connector_sharepoint}
\title{Log Read Operation for Sharepoint connector}
\usage{
\method{log_read_connector}{Connector_sharepoint}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The Connector_sharepoint object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_read_connector function for the
Connector_sharepoint class.
}
