% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_DatabricksVolume_cnt.R
\name{log_read_connector.ConnectorDatabricksVolume}
\alias{log_read_connector.ConnectorDatabricksVolume}
\title{Log Read Operation for ConnectorDatabricksVolume connector}
\usage{
\method{log_read_connector}{ConnectorDatabricksVolume}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The ConnectorDatabricksVolume object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_read_connector function for the
ConnectorDatabricksVolume class.
}
