% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_read_connector.default}
\alias{log_read_connector.default}
\title{Default log read operation}
\usage{
\method{log_read_connector}{default}(connector_object, name, ...)
}
\description{
Default log read operation
}
