% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_read_connector}
\alias{log_read_connector}
\title{Log read operation for a connector}
\usage{
log_read_connector(connector_object, name, ...)
}
\description{
Log read operation for a connector
}
