% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector.connector_fs}
\alias{log_remove_connector.connector_fs}
\title{Log remove operation for FS connector}
\usage{
\method{log_remove_connector}{connector_fs}(connector_object, name, ...)
}
\description{
Log remove operation for FS connector
}
