% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector.Connector_sharepoint}
\alias{log_remove_connector.Connector_sharepoint}
\title{Log remove operation for Sharepoint connector}
\usage{
\method{log_remove_connector}{Connector_sharepoint}(connector_object, name, ...)
}
\description{
Log remove operation for Sharepoint connector
}
