% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector.ConnectorDatabricksVolume}
\alias{log_remove_connector.ConnectorDatabricksVolume}
\title{Log remove operation for ConnectorDatabricksVolume connector}
\usage{
\method{log_remove_connector}{ConnectorDatabricksVolume}(connector_object, name, ...)
}
\description{
Log remove operation for ConnectorDatabricksVolume connector
}
