% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector.default}
\alias{log_remove_connector.default}
\title{Default log remove operation}
\usage{
\method{log_remove_connector}{default}(connector_object, name, ...)
}
\description{
Default log remove operation
}
