% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_remove_connector}
\alias{log_remove_connector}
\title{Log remove operation for a connector}
\usage{
log_remove_connector(connector_object, name, ...)
}
\description{
Log remove operation for a connector
}
