% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_write_connector.connector_dbi}
\alias{log_write_connector.connector_dbi}
\title{Log write operation for connector dbi}
\usage{
\method{log_write_connector}{connector_dbi}(connector_object, name, ...)
}
\description{
Log write operation for connector dbi
}
