% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_write_connector.ConnectorDatabricksVolume}
\alias{log_write_connector.ConnectorDatabricksVolume}
\title{Log write operation for ConnectorDatabricksVolume connector}
\usage{
\method{log_write_connector}{ConnectorDatabricksVolume}(connector_object, name, ...)
}
\description{
Log write operation for ConnectorDatabricksVolume connector
}
