% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_write_connector.default}
\alias{log_write_connector.default}
\title{Default log write operation}
\usage{
\method{log_write_connector}{default}(connector_object, name, ...)
}
\description{
Default log write operation
}
