% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_write_connector}
\alias{log_write_connector}
\title{Log write operation for a connector}
\usage{
log_write_connector(connector_object, name, ...)
}
\description{
Log write operation for a connector
}
