% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{print.connector_logger}
\alias{print.connector_logger}
\title{Print Connector Logger}
\usage{
\method{print}{connector_logger}(x, ...)
}
\arguments{
\item{x}{The connector logger object}

\item{...}{Additional arguments}
}
\value{
The result of the print operation
}
\description{
This function prints the connector logger.
}
