% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{read_cnt.connector_logger}
\alias{read_cnt.connector_logger}
\title{Log read operation for connector_logger class}
\usage{
\method{read_cnt}{connector_logger}(connector_object, name, ...)
}
\description{
Log read operation for connector_logger class
}
