% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{remove_cnt.connector_logger}
\alias{remove_cnt.connector_logger}
\title{Log remove operation for connector_logger class}
\usage{
\method{remove_cnt}{connector_logger}(connector_object, name, ...)
}
\description{
Log remove operation for connector_logger class
}
