% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{remove_cnt}
\alias{remove_cnt}
\title{Remove content from the connector}
\usage{
remove_cnt(connector_object, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition remove methods for databricks connectors implemented for
\code{\link[connector:remove_cnt]{connector::remove_cnt()}}:
}
