% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{write_cnt.connector_logger}
\alias{write_cnt.connector_logger}
\title{Log write operation for connector_logger class}
\usage{
\method{write_cnt}{connector_logger}(connector_object, x, name, ...)
}
\description{
Log write operation for connector_logger class
}
