% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector_imports.R
\name{write_cnt}
\alias{write_cnt}
\title{Write content to the connector}
\usage{
write_cnt(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{\link[connector]{connector} The connector object to use.}

\item{x}{The object to write to the connection}

\item{name}{\link{character} Name of the content to read, write, or remove. Typically the table name.}

\item{...}{Additional arguments passed to the method for the individual connector.}
}
\value{
\link{invisible} connector_object.
}
\description{
Addition write methods for databricks connectors implemented for
\code{\link[connector:write_cnt]{connector::write_cnt()}}:
}
