% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connector.logger-package.R
\docType{package}
\name{connector.logger-package}
\alias{connector.logger}
\alias{connector.logger-package}
\title{connector.logger: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NN-OpenSource/connector.logger}
  \item \url{https://miniature-couscous-2kqngew.pages.github.io/}
  \item Report bugs at \url{https://github.com/NN-OpenSource/connector.logger/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
