% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\docType{data}
\name{connector_logger}
\alias{connector_logger}
\title{Create a New Connector Logger}
\format{
An object of class \code{connector_logger} of length 0.
}
\usage{
connector_logger
}
\value{
An S3 object of class "connector_logger" containing:
\itemize{
\item An empty list
\item Class attribute set to "connector_logger"
}
}
\description{
Creates a new empty connector logger object of class
"connector_logger".
This is an S3 class constructor that initializes a logging structure for
connector operations.
}
\details{
Create a New Connector Logger
}
\examples{
logger <- connector_logger
class(logger) # Returns "connector_logger"
str(logger) # Shows empty list with class attribute

}
\keyword{datasets}
