% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{log_read_connector.Connector_sharepoint}
\alias{log_read_connector.Connector_sharepoint}
\title{Log read operation for Sharepoint connector}
\usage{
\method{log_read_connector}{Connector_sharepoint}(connector_object, name, ...)
}
\description{
Log read operation for Sharepoint connector
}
