% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_databricks_dbi_cnt.R
\name{log_write_connector.connector_databricks_dbi}
\alias{log_write_connector.connector_databricks_dbi}
\title{Log Write Operation for Databricks dbi connector}
\usage{
\method{log_write_connector}{connector_databricks_dbi}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector_databricks_dbi object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_write_connector function for the
connector_databricks_dbi class.
}
