% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_fs_cnt.R
\name{log_write_connector.connector_fs}
\alias{log_write_connector.connector_fs}
\title{Log Write Operation for FS connector}
\usage{
\method{log_write_connector}{connector_fs}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector_fs object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\description{
Implementation of the log_write_connector function for the connector_fs class.
}
