% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_add_logs.R, R/cnt_logger_generics.R
\name{print.connector_logger}
\alias{print.connector_logger}
\title{Print Method for connector_logger objects}
\usage{
\method{print}{connector_logger}(x, ...)

\method{print}{connector_logger}(x, ...)
}
\arguments{
\item{x}{The connector logger object}

\item{...}{Additional arguments}
}
\value{
The result of the next method in the dispatch chain.

The result of the print operation
}
\description{
This function prints the connector logger.
}
\details{
This method is designed to be called automatically when \code{print()} is used
on an object of class "connector_logger". It uses \code{NextMethod()} to call
the next appropriate method in the method dispatch chain, allowing for
the default or any other custom print behavior to be executed.
}
\seealso{
\code{\link{print}}
}
