% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{read_cnt.connector_logger}
\alias{read_cnt.connector_logger}
\title{Log Read Operation for connector_logger class}
\usage{
\method{read_cnt}{connector_logger}(connector_object, name, ...)
}
\arguments{
\item{connector_object}{The connector_logger object.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\value{
The result of the read operation.
}
\description{
Implementation of the log_read_connector function for the connector_logger
class.
}
