% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnt_logger_generics.R
\name{write_cnt.connector_logger}
\alias{write_cnt.connector_logger}
\title{Log Write Operation for connector_logger class}
\usage{
\method{write_cnt}{connector_logger}(connector_object, x, name, ...)
}
\arguments{
\item{connector_object}{The connector_logger object.}

\item{x}{The data to write.}

\item{name}{The name of the connector.}

\item{...}{Additional parameters.}
}
\value{
Invisible result of the write operation.
}
\description{
Implementation of the log_write_connector function for the connector_logger
class.
}
